requirements.txt
attrs==25.1.0
beautifulsoup4==4.12.3
behave==1.2.6
cached-property==2.0.1
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
cryptography==43.0.3
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
et_xmlfile==2.0.0
exchangelib==5.5.0
fire==0.7.0
flake8==7.1.1
fonttools==4.55.7
fpdf2==2.7.5
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
holidays==0.45
htmldocx==0.0.6
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isodate==0.7.2
java-access-bridge-wrapper==1.2.0
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
mccabe==0.7.0
more-itertools==10.6.0
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.5
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pdfminer.six==20221105
pendulum==3.0.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
psutil==5.9.8
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pynput-robocorp-fork==5.0.0
pyotp==2.9.0
pypdf==3.17.4
pyperclip==1.9.0
PySocks==1.7.1
pyspnego==0.11.2
pytest==8.3.4
pytest-rerunfailures==14.0
pytest-trio==0.8.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinauto==0.6.9
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
robocorp-storage==1.0.5
robotframework==6.1.1
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==6.7.0
robotframework-seleniumtestability==2.1.0
rpaframework==29.0.0
rpaframework-core==11.3.4
rpaframework-pdf==7.3.3
rpaframework-windows==8.0.0
rpds-py==0.22.3
selenium==4.15.2
simple-salesforce==1.12.6
six==1.17.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sspilib==0.2.0
stringcase==1.2.0
tenacity==8.5.0
termcolor==2.5.0
time-machine==2.16.0
trio==0.27.0
trio-websocket==0.11.1
tweepy==3.10.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.3.1
uiautomation==2.0.20
urllib3==2.3.0
validators==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.3.1
zipp==3.21.0
